split InterfaceBoard.js (extracted supporting classes to separate files)
