+++ src/Game/boardInterface/InterfaceBoard.js
@@ -1,5 +1,6 @@
 import "./board.css";
 import BoardState from "../gameLogic/history/BoardState"
+import InterfaceField from "./InterfaceField"
 
 export default class InterfaceBoard{
     /**
@@ -185,177 +186,5 @@ export default class InterfaceBoard{
             }
         }
         return position;
-    }
-}
-
-/**
- * Supporting class for the InterfaceBoard - this class and its methods should never be accessed from outside InterfaceBoard or itself.
- */
-class InterfaceField {
-    /**
-     * Constructor creating a single field of the board, with specified coordinates
-     * @param {Object} coords  - coordinates in the {x:0,y:0} object format
-     * @param {Function} onClick - the function to be called with field's coordinates ({x:0,y:0} object) when the field is clicked
-     */
-    constructor(coords, onClick) {
-        let fieldColor = (coords.x + coords.y) % 2 ? "white" : "black";
-        let fieldName = InterfaceBoard.fieldNameFromCoords(coords);
-
-        let node = document.createElement("div");
-        this._node = node;
-        this._coords = coords;
-        this._onClick = onClick;
-
-        node.id = fieldName;
-        node.classList.add("field");
-        node.classList.add('_' + fieldName[0]);
-        node.classList.add('_' + fieldName[1]);
-        node.classList.add(fieldColor);
-
-        node.onclick = function(){
-            this._onClick(this._coords);
-        }.bind(this);
-
-        this._currentPiece = null;
-    }
-
-    get node(){
-        return this._node;
-    }
-    
-    get currentPiece(){
-        return this._currentPiece;
-    }
-
-    isEmpty(){
-        return !this._currentPiece;
-    }
-
-    toggleHighlight(){
-        this._node.classList.toggle("highlighted");
-    }
-
-    applyHighlight(){
-        this._node.classList.add("highlighted");
-    }
-
-    removeHighlight(){
-        this._node.classList.remove("highlighted");
-    }
-
-    /**
-     * Removes the current piece from the field. Logical removal is immediate, actual removal from DOM is delayed by the time of exit transition
-     */
-    removePiece(){
-        if(this._currentPiece){
-            let removedPieceNode = this._currentPiece.node;
-            this._currentPiece = null;
-
-            if(removedPieceNode){
-                removedPieceNode.classList.add("moving");
-                const removedPieceStyle = getComputedStyle(removedPieceNode);
-                const animationWindow = parseFloat(removedPieceStyle.transitionDuration);
-
-                if(animationWindow>0){
-                    setTimeout(function(){
-                        removedPieceNode.remove();
-                    },animationWindow*1000);
-                }
-                else{
-                    removedPieceNode.remove();
-                }
-            }
-        }
-    }
-
-    /**
-     * Sets the field to contain the piece specified. If a piece is already present, it gets replaced.
-     * @param {String} piece type of the piece
-     * @param {String} color color of the piece
-     */
-    placePiece(piece="pawn", color="white"){
-        let newPiece = new InterfacePiece(piece,color,null);
-        if(!newPiece.equals(this._currentPiece)){
-            this.removePiece();
-            this._currentPiece = newPiece;
-            this._node.appendChild(this._currentPiece.node);
-        }
-    }
-}
-
-/**
- * Supporting class for the InterfaceBoard - this class and its methods should never be accessed from outside InterfaceBoard or itself.
- */
-class InterfacePiece{
-    constructor(piece, color){
-        this._piece = piece;
-        this._color = color;
-        this._node = null;
-    }
-
-    get piece(){
-        return this._piece;
-    }
-    get color(){
-        return this._color;
-    }
-
-    /**
-     * Returns the <div> node representing the piece. The node is created on first call.
-     */
-    get node(){
-        if(!this._node){
-            this._setUpNode();
-        }
-        return this._node;
-    }
-    
-    /**
-     * Simple method comparing whether the argument piece is logically equal (is the same kind and color) to the one it's being compared to.
-     * This method does NOT check position and node bindings - two white pawns will always be equal regardless of where they are.
-     * @param {InterfacePiece} otherPiece the piece to be compared
-     */
-    equals(otherPiece){
-        return otherPiece && this.piece===otherPiece.piece && this.color===otherPiece.color;
-    }
-
-    /**
-     * Helper method creating the visual representation of the piece as a <div> node with appropriate classes for CSS styling
-     */
-    _setUpNode(){
-        let pieceNode = document.createElement("div");
-        pieceNode.classList.add("piece");
-
-        switch(this._piece){
-            case "king":
-                pieceNode.classList.add("k");
-                break;
-            case "queen":
-                pieceNode.classList.add("q");
-                break;
-            case "rook":
-                pieceNode.classList.add("r");
-                break;
-            case "bishop":
-                pieceNode.classList.add("b");
-                break;
-            case "knight":
-                pieceNode.classList.add("kn");
-                break;
-            case "pawn":
-                pieceNode.classList.add("p");
-                break;
-            default:
-                pieceNode.classList.add("p");
-        }
-
-        if(this._color==="black"){
-            pieceNode.classList.add("bl");
-        }
-        else{
-            pieceNode.classList.add("wh");
-        }
-
-        this._node = pieceNode;
     }
 }
